+++ db/migrate/20210614155407_add_beis_identifier_to_transfers.rb
@@ -1,0 +1,6 @@
+class AddBeisIdentifierToTransfers < ActiveRecord::Migration[6.1]
+  def change
+    add_column :incoming_transfers, :beis_identifier, :string
+    add_column :outgoing_transfers, :beis_identifier, :string
+  end
+end